fix warning from conversion int to float

+++ Pepper/src/Pepper/ImGui/ImGuiLayer.cpp
@@ -59,8 +59,8 @@ void Pepper::ImGuiLayer::End() const
 {
   ImGuiIO& io = ImGui::GetIO();
   Application& app = Application::Get();
-  io.DisplaySize =
-    ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+  io.DisplaySize = ImVec2(static_cast<float>(app.GetWindow().GetWidth()),
+                          static_cast<float>(app.GetWindow().GetHeight()));
 
   // Rendering
   ImGui::Render();
